feat(versioning/ubuntu): Support for Ubuntu codenames
debian distro json
